feat: Add angle normalization
From the project instructions:
>>> In C++, function `atan2()` returns values between -pi and pi.
When calculating phi (theta in the code) in "y = z - h(x)" for radar
measurements, the resulting angle phi in the y vector should be adjusted
so that it is between -pi and pi. The Kalman filter is expecting small
angle values between the range -pi and pi. <<<